1.5.0
- Added Cinematic and Sensitivity checks
- Marked AimAssist A as experimental
- Improved Velocity A check
- Minor formatting changes
- Doubles are now floats in RotationEvent
- Fixed a Pledge false positive with blocking
- Added timestamps for when a player is given velocity and takes the proper amount
- Added ToString to CustomLocation
- Added getGcd and getMode functions to MathUtil
- Increased default reach threshold
- Removed mention of Flight E in config.yml

+++ src/main/java/net/foulest/vulture/event/RotationEvent.java
@@ -20,7 +20,7 @@ public class RotationEvent {
      *
      * @return The change in yaw.
      */
-    public double getDeltaYaw() {
+    public float getDeltaYaw() {
         return Math.abs(to.getYaw() - from.getYaw());
     }
 
@@ -29,7 +29,7 @@ public class RotationEvent {
      *
      * @return The change in pitch.
      */
-    public double getDeltaPitch() {
+    public float getDeltaPitch() {
         return Math.abs(to.getPitch() - from.getPitch());
     }
